+++ crates/horae-gui/src-tauri/src/main.rs
@@ -1,6 +1,0 @@
-// Prevents an additional console window on Windows in release.
-#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-
-fn main() {
-    horae_gui::run()
-}